+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# master (not yet released)
+**Breaking Change:**
+- Requires `kubeclient` 3.1.2 or 4.x
+
+**Changes:**
+- Add Appraisal for testing with kubeclient 3.1.2 and 4.x
+
 # v0.9.0
 - Update to not pollute the job class with our methods